Use native pointer-sized indices

+++ cpp/include/seqwin/graph.hpp
@@ -29,9 +29,9 @@ struct Node {
     /** Hash value of the minimizers represented by this node. */
     std::uint64_t hash;
     /** Start of the half-open range for this node's minimizer entries. */
-    std::uint64_t start;
+    std::size_t start;
     /** End of the half-open range for this node's minimizer entries. */
-    std::uint64_t stop;
+    std::size_t stop;
     /** Number of target assemblies containing this minimizer hash. */
     std::uint32_t n_tar;
     /** Number of non-target assemblies containing this minimizer hash. */
@@ -49,7 +49,7 @@ struct Edge {
     /** Larger endpoint hash of the undirected edge. */
     std::uint64_t second;
     /** Number of assemblies where the endpoints are adjacent. */
-    std::uint64_t weight;
+    std::size_t weight;
 };
 
 /**
@@ -59,13 +59,13 @@ struct Graph {
     /** Minimizer occurrences in all assemblies, grouped and sorted by hash. */
     NoInitArray<Kmer> kmers;
     /** Parallel to `kmers` and stores each minimizer's original generation index, ordered by genomic position. */
-    NoInitArray<std::uint64_t> idx;
+    NoInitArray<std::size_t> idx;
     /** Sorted by hash. */
     NoInitArray<Node> nodes;
     /** Sorted by hash. */
     NoInitArray<Edge> edges;
     /** Cumulative global FASTA record offsets by assembly. */
-    std::vector<std::uint64_t> record_offsets;
+    std::vector<std::size_t> record_offsets;
     /** FASTA record IDs of each assembly. */
     std::vector<std::vector<std::string>> ids_by_assembly;
 };
